streamlit/requirements.txt
streamlit
pandas
psycopg2-binary
numpy
